Add immutable.Map `updatedWith` and organize Map tests
